added first version of jinja2 templating for stdout
