+++ all.out
@@ -1,0 +1,22 @@
+Batch: HGC538
+The array seems empty. Try changing the cut_param parameter.
+The array seems empty. Try changing the cut_param parameter.
+batch HGC538:
+
+#vdp_poly   vdp_poly_rev   vdp_N   vdp_N_rev   vdp_pstop   vdp_pstop_rev   t_line_n   t_line_pstop2   t_line_pstop4   r_contact_n   r_contact_poly   fet   vdp_met_clover   vdp_met_clover_rev   vdp_cross_bridge   vdp_cross_bridge_rev   t_line_cb   v_bd   i600   v_fd   rho   d_conc   v_fb   t_ox   n_ox   c_acc   i_surf   i_surf05   i_bulk05   vdp_bulk   vdp_bulk_rev   meander_poly   meander_metal   contact_poly   contact_p   contact_n   cap   cap_tox   
+#kOhm/sq   kOhm/sq   Ohm/sq   Ohm/sq   kOhm/sq   kOhm/sq   um   um   um   Ohm   kOhm   V   mOhm/sq   mOhm/sq   kOhm/sq   kOhm/sq   um   V   uA   V   kOhm cm   1E12cm^-3   V   um   1E10cm^-3   pF   pA   pA   pA   kOhm/sq   kOhm/sq   MOhm   Ohm   MOhm   kOhm   kOhm   pF   nm   
+0.03       0.04       41.4       41.1       21.8       21.6       34.6        3.9       61.7       35.1        0.0       2.58       27.3       26.2       1.42       1.42       34.5        nan       137.8       117.0       3.86       3.60       2.50       0.73        9.4       79.1       36.36       49.6       3.59       58.8       58.6        nan       224.4        nan       415.3        nan       2.14       272.36       
+ nan       2194.67       41.1       41.0       21.7       21.8       -8458.8        1.9       61.5       34.6        nan       2.62       27.5       27.9       1.41       1.43       34.0        nan       133.1       144.6       2.70       5.15       2.54       0.73        9.6       78.9       37.16       51.5       3.84       57.3       57.9       133838.38       228.0        nan       519.1        nan       0.20       2893.05       
+ nan        nan       41.4       41.0       21.4       21.6       34.4        1.8       64.8       35.7        nan       2.40       28.1       25.8       1.41       1.42       34.3        nan       139.9       140.8       3.02       4.59       2.53       0.72        9.6       79.2       36.27       47.9       2.72       55.3       55.7        nan       223.7        nan       454.6        nan       2.13       273.88       
+ nan        nan       41.3       41.5       21.2       21.4       34.4        2.5       62.4       -7611.2        nan       2.63       26.6       27.9       1.42       1.40       34.5        nan       128.1       65.0       -41.91       -0.33       2.57       0.73        9.7       79.0       37.76       51.9       3.91       55.9       55.9        nan        nan        nan       494.4        nan       0.42       1395.37       
+ nan        nan       41.3       41.0       23.9       24.2       34.4        3.7       70.0       34.7        nan       2.40       28.9       26.7       1.41       1.41       34.5        nan       136.4       300.6       3.16       4.40       2.57       0.72        9.7       79.3       37.54       51.2       3.73       62.5       61.8        nan       224.6        nan       439.9        nan       1.95       299.58       
+ nan        nan       41.1       41.1       20.3       21.0       34.3        1.7       58.2       -22424.0        nan       2.47       26.9       27.4       1.42       1.42       34.5        nan       119.6       297.4       2.92       4.75       2.61       0.73        9.8       79.2       39.54       53.5       3.85       62.7       12547.4        nan       231.1        nan        nan        nan       2.16       269.76       
+2.33       2318.40       34.9       35.0       18.1       18.1       34.1       10.5       52.2       28.5       123.2       4.31       27.4       26.7       1.25       1.25       33.7        nan       25.6       306.7       2.75       5.06       2.39       0.67        9.9       85.6       6.62        9.4       0.61       53.9       54.4        nan       172.6       34.1       63.8       69.0       2.51       232.63       
+2.32       2.32       35.0       35.2       18.4       18.5       34.1       25.7       53.6       29.2       95.1       4.23       27.4       25.9       1.25       1.25       33.7        nan       35.9        nan        nan        nan       2.32       0.68        9.6       85.0       6.46        9.3       0.89       54.5       54.5        nan       253.3       33.6       64.6       64.6       2.28       256.17       
+2.37       2.36       35.2       35.1       18.7       18.7       34.2       33.9       52.7       29.1       113.9       4.16       28.0       26.8       1.22       1.22       33.6        nan       64.9       316.8       2.65       5.24       2.35       0.68        9.6       83.9       6.08       10.1       0.55       53.8       54.1        nan       266.2       31.1       63.3       65.2       2.49       234.14       
+2.36       2.36       35.1       35.1       18.4       18.1       34.2       10.8       52.8       28.9       127.8       4.33       27.7       26.4       1.21       1.21       33.7        nan       35.0       311.6       3.18       4.37       2.31       0.67        9.6       85.5       6.52        9.2       0.81       54.5       54.3       1.85       255.4       44.2       63.0       61.7       2.51       232.48       
+2.33       2.31       35.0       34.9       18.0       17.9       -18.5       34.5       52.1       28.4       111.1       -10.00       27.7       25.8       1.26       1.26       33.9        nan       66.1       312.3       2.91       4.78       2.41       0.67       10.0       85.6       6.18        8.8       0.48       54.2       53.3        nan       259.1       35.6       63.7       71.9       0.19       3148.65       
+2.33       2.31       35.0       35.2       18.4       18.3       34.2        0.8       52.0       29.4       111.7       4.27       27.5       26.6       1.26       1.26       33.6        nan       156.8       324.6       2.37       5.87       -1.00       -1.00       -0.0        0.4       5.82        8.5       0.81       53.4       53.6        nan       250.3       35.4       65.3       68.5       2.47       236.54       
+2.36       2.36       35.2       35.2       18.9       18.9       33.7       33.7       52.6       28.4       81.4       4.24       27.2       26.1       1.26       1.25       33.7        nan       57.7       308.4       2.72       5.11       2.38       0.69        9.5       82.9       5.59        nan        nan       53.8       54.1        nan       263.8       29.4       63.5       69.9       2.48       234.81       
+2.36       2.36       35.1       35.1       18.3       18.4       34.0       32.0       51.6       29.0       141.0       2.87       26.6       27.0       1.24       1.24       33.9        nan       52.7       312.3       3.20       4.34       2.34       0.68        9.6       84.5       6.11        8.7       0.53       53.8       54.1        nan        nan       42.6       63.2       65.4       2.51       232.02       
+
