+++ scripts/templates/all.txt
@@ -1,0 +1,14 @@
+batch {{ batch }}:
+{# this saves all pqc values to one big table  #}
+
+#{% for key in dataseries %}{% if not key.startswith('x'):%}{{ dataseries[key].name }}   {% endif %}{% endfor %}
+
+#{% for key in dataseries %}{% if not key.startswith('x'):%}{{ dataseries[key].unit }}   {% endif %}{% endfor %}
+
+{% for i in range(0, dataseries['vdp_pstop_r'].len() ) %}
+{% for key in dataseries %}
+{% if not key.startswith('x'):%}{{ dataseries[key].getValueString(i) }}       {% endif %}
+{% endfor %}
+
+{% endfor %}
+
