+++ scripts/templates/stdout.txt
@@ -1,0 +1,24 @@
+batch {{ batch }}:
+
+{% set flute1=['v_th', 'v_fb2', 'c_acc_m', 't_ox', 'n_ox', 'vdp_poly_f', 'vdp_poly_r', 'vdp_n_f', 'vdp_n_r', 'vdp_pstop_f', 'vdp_pstop_r', 'cap_l', 'cap_r'] %}
+{% set flute2=['i_surf', 't_line_n', 't_line_pstop4', 't_line_pstop2', 'v_bd'] %}
+{% set flute3=['i600', 'v_fd', 'rho', 'conc', 'meander_metal', 'vdp_metclo_f', 'vdp_metclo_r', 'vdp_p_cross_bridge_f', 'vdp_p_cross_bridge_r', 't_line_p_cross_bridge', 'vdp_bulk_f', 'vdp_bulk_r'] %}
+{% set flute4=['i_surf05', 'r_contact_n', 'contact_p', 'contact_n'] %}
+{% set labelwidth='{:40s}' %}
+{% set valuewidth='{:>12s}' %}
+{% set tables=[flute1, flute2, flute3, flute4] %}
+
+{% for table in tables %}
+
+
+{{ labelwidth.format('#') }}{% for key in table %}{{ valuewidth.format(dataseries[key].name) }}{% endfor %}
+
+{{ labelwidth.format('#') }}{% for key in table %}{{ valuewidth.format(dataseries[key].unit) }}{% endfor %}
+
+{% for i in range(0, dataseries['vdp_pstop_r'].len() ) %}
+{{ labelwidth.format(dataseries['xlabels'][i]) }}{% for key in table %}{{ valuewidth.format(dataseries[key].getValueString(i)) }}{% endfor %}
+
+{% endfor %}
+
+
+{% endfor %}
